requirements.txt
 # Production Dependencies
discord.py==2.3.2           # bot listener
sqlalchemy==2.0.25          # database ORM
psycopg2-binary             # connecting to PostgreSQL/Supabase
openai                      # generating embeddings
pgvector                    # Python client for pgvector
pydantic                    # validation (MANDATE 5.4)
python-dotenv               # local testing
scikit-learn                # semantic clustering (K-means, cosine similarity)
